+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Fix minisite migration.
 
 
 0.2 (2014-12-04)